requirements.txt
flask>=3.0.0
llama-index>=0.12.0
llama-index-vector-stores-milvus>=0.2.2
llama-index-vector-stores-elasticsearch>=0.2.0
pgvector>=0.4.1
pymilvus>=2.3.0
pymupdf>=1.25.0
sentence-transformers>=4.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
numpy>=1.21.0
sqlalchemy>=2.0.0
requests>=2.32.0
llama-index-postprocessor-dashscope-rerank==0.3.0
pydantic>=2.0.0
elasticsearch>=8.0.0